(1015) Céation de la méthode `update` du service `plan`
